refactor: 의존성 주입 경로 수정
